refactor of copy from PoC
